\name{speedRlibs}
\alias{speedRlibs}
\title{
   Package containing the required libraries (jars) for speedR
}
\description{
  speedRlibs is a container for the required libraries (jars) by speedR.
  These libraries are packaged separately, because it is approximately 20 MB and will not be so often updated like speedR.
  Thus it reduces the download time after the first installation.  
}

